implemented a way to have the version string updated with new git commits:
This is now using the latest tag in the git repository and then the number of commits afterwards and the unique specifier based on the git hash.

Evyerytime the programm is executed and it gets its version from git, it will be also write a file _version.py which includes the version string. This is needed for distribution. since the version string method obvioulsy does not work without a git repository.