Starting Refactoring of the map model
I do not like the representations of the grid and rois as dictionaries and lists. This gets very convoluted and it is often not clear what is what. Therefore we will add separate classes for MapPoint and Roi. This makes the MapModel more in line with the other models phasemodel etc.