Update cif.py
This change is to fix an error for wrongly displayed diffractograms for rhombohedral spacegroups in the hexagonal setting. 
This was previously fixed for space group number 167, by adding an if statement checking if unit cell parameters a=c. 
However, the other rhombohedral spacegroups can also be in the hexagonal setting, so by checking if a=c for all rhombohedral groups, the symmetry will be correctly assigned if the spacegroup is in the hexagonal setting.